6.NLP/patient-reply-classifier/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
comm==0.1.3
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.101.1
firebase-admin==6.2.0
fonttools==4.41.0
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
matplotlib==3.7.2
matplotlib-inline==0.1.6
msgpack==1.0.5
multidict==6.0.4
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.25.1
openai==0.27.8
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.9.1
plotly==5.15.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
regex==2023.8.8
requests==2.31.0
rsa==4.9
scikit-learn==1.3.0
scikit-plot==0.3.7
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.4.0
tornado==6.3.2
tqdm==4.66.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
wcwidth==0.2.6
xgboost==1.7.6
yarl==1.9.2
google-cloud-storage==2.10.0
requests==2.31.0
